Add Department entity and establish relationship with Employee
Introduced a new `Department` entity and updated the `Employee`
class to include a `Department` property. Updated the
`Instrument_Spec_ManagerEFCoreDbContext` to include a `DbSet<Department>`.
Created a new migration `20240803150619_AddDepartment` to add the
`Departments` table and establish a foreign key relationship with
the `Employees` table. The migration also includes an index on
the `DepartmentID` column in the `Employees` table. Updated
model snapshot to reflect these changes.