Add ConditionalAppearance attributes to DemoTask class
Added the `DevExpress.ExpressApp.ConditionalAppearance` namespace to `DemoTask.cs`. Applied an `Appearance` attribute to the `DemoTask` class to change the font color to red for `ListView` items where `Status` is not 'Completed'. Added another `Appearance` attribute to the `Priority` property to change the background color to pink (RGB: 255, 240, 240) when `Priority` equals 2.